refactor: split dashboard into components
