+++ frontend/src/components/dashboard/ScoreStreakSummary.jsx
@@ -1,0 +1,12 @@
+function ScoreStreakSummary() {
+  return (
+    <div>
+      <h3>Score / Streak</h3>
+      <p>Points: 0</p>
+      <p>Current streak: 0</p>
+      <p>Best session streak: 0</p>
+    </div>
+  );
+}
+
+export default ScoreStreakSummary;
